Junção de funcoes em um arquivo
Tudo do backend no mesmo arquivo pra evitar problemas neste momento
