--- backend/components/service_usuarios/package.json
@@ -1,6 +1,0 @@
-{
-    "name": "service_usuarios",
-    "version": "0.0.1",
-    "main": "service_usuarios.js"
-}
-
