+++ backend/server.js
@@ -1,7 +1,5 @@
 
-// Inicio da parte do banco com MongoDB
 
-//import {ControlUsuarios} from './components/usuarios'
 
 const URL_USUARIOS_GET_ALL = "/usuarios";
 const URL_USUARIOS_NOVO = "/usuarios/cadastrar/:username/:password";
@@ -13,13 +11,7 @@ const URL_USUARIOS_ALTERAR_PASSWORD = "/usuarios/alterar/password/:username/:pas
 const URL_POSTAGENS_GET_ALL = "/postagens";
 const URL_POSTAGENS_NOVO = "/postagens/novo";
 
-// Criei um "modulo" node para deixar funcionalidades em arquivo separado.
-const {ControlUsuarios, ControlPostagens} = require('service_usuarios');
-
-let controlUsuarios = new ControlUsuarios();
-let controlPostagens = new ControlPostagens();
-
-// Coisas do banco de dados. Nao funcional ainda. Nao utilizar!
+// Coisas do banco de dados mongoDB. Nao funcional ainda.
 const {MongoClient} = require('mongodb')
 const DB_NAME = 'backenddb'
 const USERS_COLLECTION_NAME = 'usuarios'
@@ -65,6 +57,172 @@ function isDadoDefined(dado) {
     }
     return status;
 }
+
+
+// Classes para manipular BD
+const MAX_POSTAGENS = 5;
+
+class ControlPostagens {
+    constructor(){
+        this.postagens = [{
+            "username": "victor",
+            "titulo": "Postagem exemplo",
+            "texto": "esta eh uma postagem!"
+        }];
+    }
+
+    buscarPorUsuario(username){
+        let postagens_filtradas = [];
+        for(const postagem of this.postagens){
+            if(postagem.username == username){
+                postagens_filtradas.push(postagem)
+            }
+        }
+        return postagens_filtradas;
+    }
+
+    novaPostagem(username, titulo, texto){
+        if(this.postagens.length >= MAX_POSTAGENS){
+            this.#deletarMaisAntigo();
+        }
+
+        const postagem = {
+            "username": username,
+            "titulo": titulo,
+            "texto": texto
+        }
+
+        this.postagens.push(postagem);
+
+        return postagem;
+    }
+
+    atualizarUsernameDasPostagens(username_antigo, username_novo){
+        for(let postagem of this.postagens){
+            if(postagem.username == username_antigo){
+                postagem.username = username_novo;
+            }
+            console.log(postagem)
+        }
+    }
+
+    #deletarMaisAntigo(){
+        this.postagens.shift();
+    }
+}
+
+class ControlUsuarios {
+    constructor(){
+        this.users = [{
+            "username": "victor",
+            "password": "abc"
+        }];
+
+    }
+
+    buscar(username) {
+        let usuario = null;
+        for(const user of this.users){
+            if(user.username == username){
+                usuario = user;
+                break;
+            }
+        }
+        return usuario;
+    }
+    
+    logar(username, password) {
+        sucesso = false;
+
+        let usuario = this.buscar(username);
+    
+        if(usuario != null){
+            if(usuario.password == password){
+                sucesso = true;
+            }
+        }
+
+        return sucesso;
+    }
+
+    cadastrar(new_username, new_password) {
+        if(!this.#isUsernameDisponivel(new_username)){
+            return null;
+        }
+
+        const novo_usuario = this.#criarObjetoUsuario(new_username, new_password);
+        this.users.push(novo_usuario);
+
+        return novo_usuario;
+    }
+
+    alterarUsername(username_atual, username_novo) {
+
+        let usuario = this.buscar(username_atual);
+        let usuario_com_nome_novo = this.buscar(username_novo);
+
+        let sucesso = false;
+
+        if((usuario != null) && (usuario_com_nome_novo == null)) {
+            usuario.username = username_novo;
+            sucesso = true;
+        }
+
+        return sucesso;
+    }
+
+    alterarPassword(username, password_atual, password_novo) {
+        let usuario = this.buscar(username);
+
+        if(usuario != null) {
+            if(usuario.password = password_atual){
+                usuario.password = password_novo;
+            }
+        }
+        return usuario
+    }
+
+    deletar(username) {
+        let usuario = this.buscar(username);
+    
+        if(usuario == null){
+            return(null);
+        }
+        else{
+            const index_usuario = this.users.findIndex(user => {
+                return user.username === username;
+            })
+            const usuario_removido = this.users.splice(index_usuario, 1);
+
+            return(usuario_removido[0]);
+        }
+    }
+
+    #isUsernameDisponivel(username) {
+        let is_disponivel = true;
+    
+        const usuario = this.buscar(username)
+
+        if(usuario != null){
+            is_disponivel = false;
+        }
+
+        return is_disponivel;
+    }
+    
+    #criarObjetoUsuario(username, password) {
+        return {
+            "username": username,
+            "password": password
+        };
+    }
+
+}
+
+let controlUsuarios = new ControlUsuarios();
+let controlPostagens = new ControlPostagens();
+
+// fim de classes para manipular db
 
 // Inicio da parte do express
